pyproject.toml
[project]
name = "adk01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=2.13.1",
    "google-adk[a2a,memory]>=1.19.0",
    "google-genai[adk,mcp]>=1.52.0",
    "uvicorn>=0.38.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.2.1",
    "a2a-sdk>=0.3.16",
]
